+++ spec/features/transfers/users_can_create_a_transfer_project_spec.rb
@@ -54,6 +54,10 @@ RSpec.feature "Users can create new conversion projects" do
       choose "No"
     end
 
+    within("#inadequate-ofsted") do
+      choose "No"
+    end
+
     fill_in "Handover comments", with: "This is a handover note."
 
     within("#provisional-transfer-date") do