+++ spec/forms/transfer/create_project_form_spec.rb
@@ -101,6 +101,14 @@ RSpec.describe Transfer::CreateProjectForm, type: :model do
         expect(form.errors.messages[:incoming_trust_ukprn]).to include I18n.t("errors.attributes.incoming_trust_ukprn.ukprns_must_not_match")
       end
     end
+
+    describe "inadequate Ofsted" do
+      it "cannot be blank" do
+        form = build(:create_transfer_project_form, inadequate_ofsted: "")
+
+        expect(form).to be_invalid
+      end
+    end
   end
 
   describe "urn" do
@@ -260,6 +268,14 @@ RSpec.describe Transfer::CreateProjectForm, type: :model do
         expect(Note.count).to eq(1)
         expect(Note.last.body).to eq("This is the handover note.")
         expect(Note.last.task_identifier).to eq("handover")
+      end
+
+      it "updates the tasks data with the other values" do
+        form = build(:create_transfer_project_form, inadequate_ofsted: true)
+        form.save
+
+        expect(Transfer::TasksData.count).to be 1
+        expect(Transfer::TasksData.first.inadequate_ofsted).to be true
       end
     end