+++ CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add ability to import GIAS groups from csv data
 - transfer project now collect if they are a result of 2RI and show this
   information on the about the project tab
+- transfer project now collect if they are a result of 2RI and show this
+  information on the about the project tab
+- transfer projects now collect if they are the result of an inadequate Ofsted
+  inspection and show this information on the about the project tab
 
 ## [Release-46][release-46]